Review comments incorporated. 1. Moved out vxlan related config and show utilities to a separate file. 2. Moved some utility functions to utilitie_common
