+++ config/main.py
@@ -23,7 +23,6 @@ from utilities_common.intf_filter import parse_interface_in_filter
 import utilities_common.cli as clicommon
 from .utils import log
 
-
 from . import aaa
 from . import chassis_modules
 from . import console
@@ -34,6 +33,7 @@ from . import mlnx
 from . import muxcable
 from . import nat
 from . import vlan
+from . import vxlan
 from .config_mgmt import ConfigMgmtDPB
 
 CONTEXT_SETTINGS = dict(help_option_names=['-h', '--help', '-?'])
@@ -341,76 +341,7 @@ def interface_name_is_valid(config_db, interface_name):
                     return True
     return False
 
-#
-# Use this method to validate unicast IPv4 address
-#
-def is_ip4_addr_valid(addr, display):
-    v4_invalid_list = [ipaddress.IPv4Address(unicode('0.0.0.0')), ipaddress.IPv4Address(unicode('255.255.255.255'))]
-    try:
-        ip = ipaddress.ip_address(unicode(addr))
-        if (ip.version == 4):
-            if (ip.is_reserved):
-                if display:
-                    click.echo ("{} Not Valid, Reason: IPv4 reserved address range.".format(addr))
-                return False
-            elif (ip.is_multicast):
-                if display:
-                    click.echo ("{} Not Valid, Reason: IPv4 Multicast address range.".format(addr))
-                return False
-            elif (ip in v4_invalid_list):
-                if display:
-                    click.echo ("{} Not Valid.".format(addr))
-                return False
-            else:
-                return True
-
-        else:
-            if display:
-                click.echo ("{} Not Valid, Reason: Not an IPv4 address".format(addr))
-            return False
-
-    except ValueError:
-        return False
-
-def vlan_id_is_valid(vid):
-    """Check if the vlan id is in acceptable range (between 1 and 4094)
-    """
-
-    if vid<1 or vid>4094:
-        return False
-
-    return True
-
-def vni_id_is_valid(vni):
-    """Check if the vni id is in acceptable range (between 1 and 2^24)
-    """
-
-    if (vni < 1) or (vni > 16777215):
-        return False
-
-    return True
-
-def is_vni_vrf_mapped(ctx, vni):
-    """Check if the vni is mapped to vrf
-    """
-
-    found = 0
-    db = ctx.obj['db']
-    vrf_table = db.get_table('VRF')
-    vrf_keys = vrf_table.keys()
-    if vrf_keys is not None:
-      for vrf_key in vrf_keys:
-        if ('vni' in vrf_table[vrf_key] and vrf_table[vrf_key]['vni'] == vni):
-           found = 1
-           break
-
-    if (found == 1):
-        print "VNI {} mapped to Vrf {}, Please remove VRF VNI mapping".format(vni, vrf_key)
-        return False
-
-    return True
-
-def interface_name_to_alias(interface_name):
+def interface_name_to_alias(config_db, interface_name):
     """Return alias interface name if default name is given as argument
     """
     # If the input parameter config_db is None, derive it from interface.
@@ -953,6 +884,7 @@ config.add_command(kube.kubernetes)
 config.add_command(muxcable.muxcable)
 config.add_command(nat.nat)
 config.add_command(vlan.vlan)
+config.add_command(vxlan.vxlan)
 
 @config.command()
 @click.option('-y', '--yes', is_flag=True, callback=_abort_if_false,
@@ -2737,7 +2669,7 @@ def add_vrf_vni_map(ctx, vrfname, vni):
     if not vni.isdigit():
         ctx.fail("Invalid VNI {}. Only valid VNI is accepted".format(vni))
 
-    if vni_id_is_valid(int(vni)) is False:
+    if clicommon.vni_id_is_valid(int(vni)) is False:
         ctx.fail("Invalid VNI {}. Valid range [1 to 16777215].".format(vni))
 
     vxlan_table = config_db.get_table('VXLAN_TUNNEL_MAP')
@@ -3752,406 +3684,5 @@ def delete(ctx):
     sflow_tbl['global'].pop('agent_id')
     config_db.set_entry('SFLOW', 'global', sflow_tbl['global'])
 
-#
-# 'feature' command ('config feature name state')
-# 
-@config.command('feature')
-@click.argument('name', metavar='<feature-name>', required=True)
-@click.argument('state', metavar='<feature-state>', required=True, type=click.Choice(["enabled", "disabled"]))
-def feature_status(name, state):
-    """ Configure status of feature"""
-    config_db = ConfigDBConnector()
-    config_db.connect()
-    status_data = config_db.get_entry('FEATURE', name)
-
-    if not status_data:
-        click.echo(" Feature '{}' doesn't exist".format(name))
-        return
-
-    config_db.mod_entry('FEATURE', name, {'status': state})
-
-#
-# 'container' group ('config container ...')
-#
-@config.group(cls=AbbreviationGroup, name='container', invoke_without_command=False)
-def container():
-    """Modify configuration of containers"""
-    pass
-
-#
-# 'feature' group ('config container feature ...')
-#
-@container.group(cls=AbbreviationGroup, name='feature', invoke_without_command=False)
-def feature():
-    """Modify configuration of container features"""
-    pass
-
-#
-# 'autorestart' subcommand ('config container feature autorestart ...')
-#
-@feature.command(name='autorestart', short_help="Configure the status of autorestart feature for specific container")
-@click.argument('container_name', metavar='<container_name>', required=True)
-@click.argument('autorestart_status', metavar='<autorestart_status>', required=True, type=click.Choice(["enabled", "disabled"]))
-def autorestart(container_name, autorestart_status):
-    config_db = ConfigDBConnector()
-    config_db.connect()
-    container_feature_table = config_db.get_table('CONTAINER_FEATURE')
-    if not container_feature_table:
-        click.echo("Unable to retrieve container feature table from Config DB.")
-        return
-
-    if not container_feature_table.has_key(container_name):
-        click.echo("Unable to retrieve features for container '{}'".format(container_name))
-        return
-
-    config_db.mod_entry('CONTAINER_FEATURE', container_name, {'auto_restart': autorestart_status})
-
-#
-# 'vxlan' group ('config vxlan ...')
-#
-@config.group()
-@click.pass_context
-def vxlan(ctx):
-    config_db = ConfigDBConnector()
-    config_db.connect()
-    ctx.obj = {'db': config_db}
-
-@vxlan.command('add')
-@click.argument('vxlan_name', metavar='<vxlan_name>', required=True)
-@click.argument('src_ip', metavar='<src_ip>', required=True)
-@click.pass_context
-def add_vxlan(ctx, vxlan_name, src_ip):
-    """Add VXLAN"""
-    if not is_ip4_addr_valid(src_ip, True):
-        ctx.fail("{} invalid src ip address".format(src_ip))  
-    db = ctx.obj['db']
-
-    vxlan_keys = db.keys('CONFIG_DB', "VXLAN_TUNNEL|*")
-    if not vxlan_keys:
-      vxlan_count = 0
-    else:
-      vxlan_count = len(vxlan_keys)
-
-    if(vxlan_count > 0):
-        ctx.fail("VTEP already configured.")  
-
-    fvs = {'src_ip': src_ip}
-    db.set_entry('VXLAN_TUNNEL', vxlan_name, fvs)
-
-@vxlan.command('del')
-@click.argument('vxlan_name', metavar='<vxlan_name>', required=True)
-@click.pass_context
-def del_vxlan(ctx, vxlan_name):
-    """Del VXLAN"""
-    db = ctx.obj['db']
-
-    vxlan_keys = db.keys('CONFIG_DB', "VXLAN_EVPN_NVO|*")
-    if not vxlan_keys:
-      vxlan_count = 0
-    else:
-      vxlan_count = len(vxlan_keys)
-
-    if(vxlan_count > 0):
-        ctx.fail("Please delete the EVPN NVO configuration.")  
-
-    vxlan_keys = db.keys('CONFIG_DB', "VXLAN_TUNNEL_MAP|*")
-    if not vxlan_keys:
-      vxlan_count = 0
-    else:
-      vxlan_count = len(vxlan_keys)
-
-    if(vxlan_count > 0):
-        ctx.fail("Please delete all VLAN VNI mappings.")  
-
-    db.set_entry('VXLAN_TUNNEL', vxlan_name, None)
-
-@vxlan.group('evpn_nvo')
-@click.pass_context
-def vxlan_evpn_nvo(ctx):
-    pass
-
-@vxlan_evpn_nvo.command('add')
-@click.argument('nvo_name', metavar='<nvo_name>', required=True)
-@click.argument('vxlan_name', metavar='<vxlan_name>', required=True)
-@click.pass_context
-def add_vxlan_evpn_nvo(ctx, nvo_name, vxlan_name):
-    """Add NVO"""
-    db = ctx.obj['db']
-    vxlan_keys = db.keys('CONFIG_DB', "VXLAN_EVPN_NVO|*")
-    if not vxlan_keys:
-      vxlan_count = 0
-    else:
-      vxlan_count = len(vxlan_keys)
-
-    if(vxlan_count > 0):
-        ctx.fail("EVPN NVO already configured")  
-
-    if len(db.get_entry('VXLAN_TUNNEL', vxlan_name)) == 0:
-        ctx.fail("VTEP {} not configured".format(vxlan_name))
-
-    fvs = {'source_vtep': vxlan_name}
-    db.set_entry('VXLAN_EVPN_NVO', nvo_name, fvs)
-
-@vxlan_evpn_nvo.command('del')
-@click.argument('nvo_name', metavar='<nvo_name>', required=True)
-@click.pass_context
-def del_vxlan_evpn_nvo(ctx, nvo_name):
-    """Del NVO"""
-    db = ctx.obj['db']
-    vxlan_keys = db.keys('CONFIG_DB', "VXLAN_TUNNEL_MAP|*")
-    if not vxlan_keys:
-      vxlan_count = 0
-    else:
-      vxlan_count = len(vxlan_keys)
-
-    if(vxlan_count > 0):
-        ctx.fail("Please delete all VLAN VNI mappings.")  
-    db.set_entry('VXLAN_EVPN_NVO', nvo_name, None)
-
-@vxlan.group('map')
-@click.pass_context
-def vxlan_map(ctx):
-    pass
-
-@vxlan_map.command('add')
-@click.argument('vxlan_name', metavar='<vxlan_name>', required=True)
-@click.argument('vlan', metavar='<vlan_id>', required=True)
-@click.argument('vni', metavar='<vni>', required=True)
-@click.pass_context
-def add_vxlan_map(ctx, vxlan_name, vlan, vni):
-    """Add VLAN-VNI map entry"""
-    if not vlan.isdigit():
-        ctx.fail("Invalid vlan {}. Only valid vlan is accepted".format(vni))
-    if vlan_id_is_valid(int(vlan)) is False:
-        ctx.fail(" Invalid Vlan Id , Valid Range : 1 to 4094 ")
-    if not vni.isdigit():
-        ctx.fail("Invalid VNI {}. Only valid VNI is accepted".format(vni))
-    if vni_id_is_valid(int(vni)) is False:
-        ctx.fail("Invalid VNI {}. Valid range [1 to 16777215].".format(vni))
-
-    db = ctx.obj['db']
-    vlan_name = "Vlan" + vlan
-
-    if len(db.get_entry('VXLAN_TUNNEL', vxlan_name)) == 0:
-        ctx.fail("VTEP {} not configured".format(vxlan_name))
-
-    if len(db.get_entry('VLAN', vlan_name)) == 0:
-        ctx.fail("{} not configured".format(vlan_name))
-
-    vxlan_table = db.get_table('VXLAN_TUNNEL_MAP')
-    vxlan_keys = vxlan_table.keys()
-    if vxlan_keys is not None:
-      for key in vxlan_keys:
-        if (vxlan_table[key]['vlan'] == vlan_name):
-           ctx.fail(" Vlan Id already mapped ")
-        if (vxlan_table[key]['vni'] == vni):
-           ctx.fail(" VNI Id already mapped ")
-
-    fvs = {'vni': vni,
-           'vlan' : vlan_name}
-    mapname = vxlan_name + '|' + 'map_' + vni + '_' + vlan_name
-    db.set_entry('VXLAN_TUNNEL_MAP', mapname, fvs)
-
-@vxlan_map.command('del')
-@click.argument('vxlan_name', metavar='<vxlan_name>', required=True)
-@click.argument('vlan', metavar='<vlan_id>', required=True)
-@click.argument('vni', metavar='<vni>', required=True)
-@click.pass_context
-def del_vxlan_map(ctx, vxlan_name, vlan, vni):
-    """Del VLAN-VNI map entry"""
-    if not vlan.isdigit():
-        ctx.fail("Invalid vlan {}. Only valid vlan is accepted".format(vni))
-    if vlan_id_is_valid(int(vlan)) is False:
-        ctx.fail(" Invalid Vlan Id , Valid Range : 1 to 4094 ")
-    if not vni.isdigit():
-        ctx.fail("Invalid VNI {}. Only valid VNI is accepted".format(vni))
-    if vni_id_is_valid(int(vni)) is False:
-        ctx.fail("Invalid VNI {}. Valid range [1 to 16777215].".format(vni))
-
-    db = ctx.obj['db']
-    if len(db.get_entry('VXLAN_TUNNEL', vxlan_name)) == 0:
-        ctx.fail("VTEP {} not configured".format(vxlan_name))
-    found = 0
-    vrf_table = db.get_table('VRF')
-    vrf_keys = vrf_table.keys()
-    if vrf_keys is not None:
-      for vrf_key in vrf_keys:
-        if ('vni' in vrf_table[vrf_key] and vrf_table[vrf_key]['vni'] == vni):
-           found = 1
-           break
-
-    if (found == 1):
-        ctx.fail("VNI mapped to vrf {}, Please remove VRF VNI mapping".format(vrf_key))
-
-    mapname = vxlan_name + '|' + 'map_' + vni + '_' + vlan
-    db.set_entry('VXLAN_TUNNEL_MAP', mapname, None)
-    mapname = vxlan_name + '|' + 'map_' + vni + '_Vlan' + vlan
-    db.set_entry('VXLAN_TUNNEL_MAP', mapname, None)
-
-@vxlan.group('map_range')
-@click.pass_context
-def vxlan_map_range(ctx):
-    pass
-
-@vxlan_map_range.command('add')
-@click.argument('vxlan_name', metavar='<vxlan_name>', required=True)
-@click.argument('vlan_start', metavar='<vlan_start>', required=True, type=int)
-@click.argument('vlan_end', metavar='<vlan_end>', required=True, type=int)
-@click.argument('vni_start', metavar='<vni_start>', required=True, type=int)
-@click.pass_context
-def add_vxlan_map_range(ctx, vxlan_name, vlan_start, vlan_end, vni_start):
-    """Add Range of vlan-vni mappings"""
-    if vlan_id_is_valid(vlan_start) is False:
-        ctx.fail(" Invalid Vlan Id , Valid Range : 1 to 4094 ")
-    if vlan_id_is_valid(vlan_end) is False:
-        ctx.fail(" Invalid Vlan Id , Valid Range : 1 to 4094 ")
-    if (vlan_start > vlan_end):
-       ctx.fail("vlan_end should be greater or equal to vlan_start")
-    if vni_id_is_valid(vni_start) is False:
-        ctx.fail("Invalid VNI {}. Valid range [1 to 16777215].".format(vni_start))
-    if vni_id_is_valid(vni_start+vlan_end-vlan_start) is False:
-        ctx.fail("Invalid VNI End {}. Valid range [1 to 16777215].".format(vni_start))
-
-    db = ctx.obj['db']
-    if len(db.get_entry('VXLAN_TUNNEL', vxlan_name)) == 0:
-        ctx.fail("VTEP {} not configured".format(vxlan_name))
-    vlan_end = vlan_end + 1
-    vxlan_table = db.get_table('VXLAN_TUNNEL_MAP')
-    vxlan_keys = vxlan_table.keys()
-
-    for vid in range (vlan_start, vlan_end):
-       vlan_name = 'Vlan{}'.format(vid)
-       vnid = vni_start+vid-vlan_start
-       vni_name = '{}'.format(vnid)
-       match_found = 'no'
-       if len(db.get_entry('VLAN', vlan_name)) == 0:
-         click.echo("{} not configured".format(vlan_name))
-         continue
-       if vxlan_keys is not None:
-          for key in vxlan_keys:
-            if (vxlan_table[key]['vlan'] == vlan_name):
-              print(vlan_name + " already mapped")
-              match_found = 'yes'
-              break
-            if (vxlan_table[key]['vni'] == vni_name):
-              print("VNI:" + vni_name + " already mapped ")
-              match_found = 'yes'
-              break
-       if (match_found == 'yes'):
-         continue
-       fvs = {'vni': vni_name,
-              'vlan' : vlan_name}
-       mapname = vxlan_name + '|' + 'map_' + vni_name + '_' + vlan_name
-       db.set_entry('VXLAN_TUNNEL_MAP', mapname, fvs)
-
-@vxlan_map_range.command('del')
-@click.argument('vxlan_name', metavar='<vxlan_name>', required=True)
-@click.argument('vlan_start', metavar='<vlan_start>', required=True, type=int)
-@click.argument('vlan_end', metavar='<vlan_end>', required=True, type=int)
-@click.argument('vni_start', metavar='<vni_start>', required=True, type=int)
-@click.pass_context
-def del_vxlan_map_range(ctx, vxlan_name, vlan_start, vlan_end, vni_start):
-    """Del Range of vlan-vni mappings"""
-    if vlan_id_is_valid(vlan_start) is False:
-        ctx.fail(" Invalid Vlan Id , Valid Range : 1 to 4094 ")
-    if vlan_id_is_valid(vlan_end) is False:
-        ctx.fail(" Invalid Vlan Id , Valid Range : 1 to 4094 ")
-    if (vlan_start > vlan_end):
-       ctx.fail("vlan_end should be greater or equal to vlan_start")
-    if vni_id_is_valid(vni_start) is False:
-        ctx.fail("Invalid VNI {}. Valid range [1 to 16777215].".format(vni_start))
-    if vni_id_is_valid(vni_start+vlan_end-vlan_start) is False:
-        ctx.fail("Invalid VNI End {}. Valid range [1 to 16777215].".format(vni_start))
-
-    db = ctx.obj['db']
-    if len(db.get_entry('VXLAN_TUNNEL', vxlan_name)) == 0:
-        ctx.fail("VTEP {} not configured".format(vxlan_name))
-
-    vlan_end = vlan_end + 1
-    for vid in range (vlan_start, vlan_end):
-       vlan_name = 'Vlan{}'.format(vid)
-       vnid = vni_start+vid-vlan_start
-       vni_name = '{}'.format(vnid)
-       if is_vni_vrf_mapped(ctx, vni_name) is False:
-           print "Skipping Vlan {} VNI {} mapped delete. ".format(vlan_name, vni_name)
-           continue
-
-       mapname = vxlan_name + '|' + 'map_' + vni_name + '_' + vlan_name
-       db.set_entry('VXLAN_TUNNEL_MAP', mapname, None)
-
-#######
-#
-# 'neigh_suppress' group ('config neigh_suppress...')
-#
-@config.group()
-@click.pass_context
-def neigh_suppress(ctx):
-    """ Neighbour Suppress VLAN-related configuration """
-    config_db = ConfigDBConnector()
-    config_db.connect()
-    ctx.obj = {'db': config_db}
-
-@neigh_suppress.command('enable')
-@click.argument('vid', metavar='<vid>', required=True, type=int)
-@click.pass_context
-def enable_neigh_suppress(ctx, vid):
-    db = ctx.obj['db']
-    if vlan_id_is_valid(vid) is False:
-        ctx.fail(" Invalid Vlan Id , Valid Range : 1 to 4094 ")
-    vlan = 'Vlan{}'.format(vid)
-    if len(db.get_entry('VLAN', vlan)) == 0:
-        click.echo("{} doesn't exist".format(vlan))
-        return
-    fvs = {'suppress': "on"}
-    db.set_entry('SUPPRESS_VLAN_NEIGH', vlan, fvs)
-
-@neigh_suppress.command('disable')
-@click.argument('vid', metavar='<vid>', required=True, type=int)
-@click.pass_context
-def disable_neigh_suppress(ctx, vid):
-    db = ctx.obj['db']
-    if vlan_id_is_valid(vid) is False:
-        ctx.fail(" Invalid Vlan Id , Valid Range : 1 to 4094 ")
-    vlan = 'Vlan{}'.format(vid)
-    db.set_entry('SUPPRESS_VLAN_NEIGH', vlan, None)
-#######
-#
-# 'neigh_suppress' group ('config neigh_suppress...')
-#
-@config.group()
-@click.pass_context
-def neigh_suppress(ctx):
-    """ Neighbour Suppress VLAN-related configuration """
-    config_db = ConfigDBConnector()
-    config_db.connect()
-    ctx.obj = {'db': config_db}
-
-@neigh_suppress.command('enable')
-@click.argument('vid', metavar='<vid>', required=True, type=int)
-@click.pass_context
-def enable_neigh_suppress(ctx, vid):
-    db = ctx.obj['db']
-    if vlan_id_is_valid(vid) is False:
-        ctx.fail(" Invalid Vlan Id , Valid Range : 1 to 4094 ")
-    vlan = 'Vlan{}'.format(vid)
-    if len(db.get_entry('VLAN', vlan)) == 0:
-        click.echo("{} doesn't exist".format(vlan))
-        return
-    fvs = {'suppress': "on"}
-    db.set_entry('SUPPRESS_VLAN_NEIGH', vlan, fvs)
-
-@neigh_suppress.command('disable')
-@click.argument('vid', metavar='<vid>', required=True, type=int)
-@click.pass_context
-def disable_neigh_suppress(ctx, vid):
-    db = ctx.obj['db']
-    if vlan_id_is_valid(vid) is False:
-        ctx.fail(" Invalid Vlan Id , Valid Range : 1 to 4094 ")
-    vlan = 'Vlan{}'.format(vid)
-    db.set_entry('SUPPRESS_VLAN_NEIGH', vlan, None)
-
-
-
 if __name__ == '__main__':
     config()
-
